+++ ChangeLog.md
@@ -1,4 +1,14 @@
-﻿### Version 3.2.0
+﻿### Version 4.1.0
+
+* Updated to Tesseract 4.1.0
+* Requires VC++ 2019 runtime
+
+### Version 4.0.0
+
+* Updated to Tesseract 4.0.0
+* Requires VC++ 2017 runtime
+
+### Version 3.2.0
 
 * Support for .Net Standard 2.0 (.net Core 2) - [Issue ]()
 * Removed support for TESSDATA environment variable